requirements.txt
h5py==2.8.0rc1
numpy==1.14.2
Pillow==5.0.0
tensorflow==1.8.0
tensorflow-tensorboard==1.5.1